created "osp-sensor-control-interface.h". extracted from osp_HostInterface.h
